move delay filter to source_utils
