+++ src/pyfhd/gridding/filters.py
@@ -2,7 +2,7 @@ import numpy as np
 from numpy.typing import NDArray
 from logging import Logger
 
-from ..pyfhd_tools.pyfhd_utils import weight_invert, spectral_window, rebin
+from ..pyfhd_tools.pyfhd_utils import weight_invert
 from . import gridding_utils
 
 
@@ -95,110 +95,3 @@ def filter_uv_uniform(
         filter_use /= np.mean(filter_use)
     # Return the filtered
     return image_uv * filter_use, filter_use
-
-
-def vis_delay_filter(
-    vis_model_arr: NDArray[np.complex128], *, obs: dict, params: dict
-) -> NDArray[np.complex128]:
-    """
-    Apply a delay space filter at the horizon to remove fft artifacts.
-
-    Model visibilities are phased to zenith, windowed, transformed to delay space,
-    cut at the horizon, transformed back to visibility space, unwindowed, unphased
-    from zenith, and frequency cut to match the data.
-
-    Parameters
-    ----------
-    vis_model_arr : ndarray of complex
-        Input model visibilities to be filtered.
-    obs : dict
-        The observation metadata dictionary.
-    params : dict
-        Visibility metadata dictionary.
-
-    Returns
-    -------
-    vis_model_arr : ndarray of complex
-        Delay filtered visibilities with half the number of frequencies as the
-        input visibilities.
-    """
-    # u,v,w are in light travel time in seconds
-    freq_arr = obs["baseline_info"]["freq"]
-    freq_res = obs["freq_res"]
-    n_pol = obs["n_pol"]
-    kbinsize = obs["kpix"]
-    nfreq = freq_arr.size
-    nbl = params["uu"].size
-
-    data = vis_model_arr.transpose(1, 2, 0)
-
-    # test with removing zeroed visibilities instead
-    total_data = np.sum(np.abs(data), axis=(0, 1))
-    bi_use = np.nonzero(total_data != 0)
-    cross_inds = np.nonzero(params["antenna1"] != params["antenna2"])
-    bi_use = np.intersect1d(bi_use, cross_inds)
-
-    data = data[:, bi_use]
-    uu = params["uu"][bi_use]
-    vv = params["vv"][bi_use]
-    ww = params["ww"][bi_use]
-    bb = np.sqrt(uu**2.0 + vv**2.0 + ww**2.0)
-    nbl = bi_use.size
-
-    # Phase to zenith -- easier calculations of the location of the horizon when
-    # phased to zenith
-    dimension = obs["dimension"]
-    dx = obs["obsx"] - obs["zenx"]
-    dy = obs["obsy"] - obs["zeny"]
-    dx *= 2.0 * np.pi / dimension
-    dy *= 2.0 * np.pi / dimension
-    phase = (
-        np.outer(freq_arr, uu) * dx / kbinsize + np.outer(freq_arr, vv) * dy / kbinsize
-    )
-    rephase_vals = np.cos(phase) + 1j * np.sin(phase)
-    rephase_vals = np.repeat((rephase_vals[:, :, np.newaxis]), n_pol, axis=2)
-
-    data *= rephase_vals
-    del uu, vv, ww
-
-    # Apply window function
-    window = spectral_window(nfreq, type="Blackman-Harris", periodic=True)
-    norm_factor = np.sqrt(nfreq / np.sum(window**2.0))
-    window = window * norm_factor
-    window_expand = rebin(window, (nfreq, nbl, n_pol), sample=True)
-    data = data * window_expand
-
-    # FFT along freq axis
-    spectra = np.fft.fftshift(np.fft.fft(data, axis=0), axes=0)
-    del data, window
-
-    # Cut at the horizon
-    tau_cut = 1.0
-
-    # Calculate upper and lower delay limits for each baseline
-    lower_limit = nfreq * (0.5 - bb * tau_cut * freq_res)
-    upper_limit = nfreq * (0.5 + bb * tau_cut * freq_res)
-
-    for freq_i in range(nfreq):
-        mask_high = freq_i / upper_limit
-        mask_high_inds = np.nonzero(mask_high > 1.0)
-        if mask_high_inds[0].size > 0:
-            spectra[freq_i, mask_high_inds] = 0
-
-        mask_low = freq_i / lower_limit
-        mask_low_inds = np.nonzero(mask_low < 1.0)
-        if mask_low_inds[0].size > 0:
-            spectra[freq_i, mask_low_inds] = 0
-
-    masked_data = np.fft.ifft(np.fft.fftshift(spectra, axes=0), axis=0)
-    masked_data = masked_data / window_expand
-
-    # Unphase from zenith and cut to the desired band
-    masked_data *= 1.0 / rephase_vals
-    masked_data = masked_data.transpose(2, 0, 1)
-    freq_ind_min = nfreq // 4
-    freq_ind_max = 3 * nfreq // 4
-    vis_model_arr = vis_model_arr[:, freq_ind_min:freq_ind_max]
-    vis_model_arr[:, :, bi_use] = masked_data[:, freq_ind_min:freq_ind_max]
-
-    return vis_model_arr
