+++ src/pyfhd/source_modeling/source_utils.py
@@ -4,7 +4,6 @@ from pathlib import Path
 
 from astropy import units
 import numpy as np
-from pyuvdata.utils.types import BoolArray, FloatArray
 from pyradiosky import SkyModel
 
 from ..beam_setup.beam_utils import beam_image
@@ -12,9 +11,11 @@ from ..pyfhd_tools.pyfhd_utils import (
     angle_difference,
     region_grow,
     resistant_mean,
+    spectral_window,
     weight_invert,
 )
 from ..pyfhd_tools.unit_conv import pixel_to_radec, radec_to_pixel
+from ..pyfhd_tools.types import BoolArray, FloatArray, ComplexArray
 
 
 def generate_source_cal_skymodel(
@@ -1107,3 +1108,111 @@ def source_dft_model(
     logger.info(f"DFT timing: {dft_timing} ({skymodel.Ncomponents} sources)")
 
     return model_uv_arr
+
+
+def vis_delay_filter(
+    vis_model_arr: ComplexArray, *, obs: dict, params: dict
+) -> ComplexArray:
+    """
+    Apply a delay space filter at the horizon to remove fft artifacts.
+
+    Model visibilities are phased to zenith, windowed, transformed to delay space,
+    cut at the horizon, transformed back to visibility space, unwindowed, unphased
+    from zenith, and frequency cut to match the data.
+
+    Parameters
+    ----------
+    vis_model_arr : ndarray of complex
+        Input model visibilities to be filtered.
+    obs : dict
+        The observation metadata dictionary.
+    params : dict
+        Visibility metadata dictionary.
+
+    Returns
+    -------
+    vis_model_arr : ndarray of complex
+        Delay filtered visibilities with half the number of frequencies as the
+        input visibilities.
+    """
+    # u,v,w are in light travel time in seconds
+    freq_arr = obs["baseline_info"]["freq"]
+    freq_res = obs["freq_res"]
+    n_pol = obs["n_pol"]
+    kbinsize = obs["kpix"]
+    nfreq = freq_arr.size
+    nbl = params["uu"].size
+
+    data = vis_model_arr.transpose(1, 2, 0)
+
+    # test with removing zeroed visibilities instead
+    total_data = np.sum(np.abs(data), axis=(0, 2))
+    bi_use = np.nonzero(total_data != 0)
+    cross_inds = np.nonzero(params["antenna1"] != params["antenna2"])
+    bi_use = np.intersect1d(bi_use, cross_inds)
+
+    data = data[:, bi_use]
+    uu = params["uu"][bi_use]
+    vv = params["vv"][bi_use]
+    ww = params["ww"][bi_use]
+    bb = np.sqrt(uu**2.0 + vv**2.0 + ww**2.0)
+    nbl = bi_use.size
+
+    # Phase to zenith -- easier calculations of the location of the horizon when
+    # phased to zenith
+    dimension = obs["dimension"]
+    dx = obs["obsx"] - obs["zenx"]
+    dy = obs["obsy"] - obs["zeny"]
+    dx *= 2.0 * np.pi / dimension
+    dy *= 2.0 * np.pi / dimension
+    phase = (
+        np.outer(freq_arr, uu) * dx / kbinsize + np.outer(freq_arr, vv) * dy / kbinsize
+    )
+    rephase_vals = np.cos(phase) + 1j * np.sin(phase)
+    rephase_vals = np.repeat((rephase_vals[:, :, np.newaxis]), n_pol, axis=2)
+    data *= rephase_vals
+    del uu, vv, ww
+
+    # Apply window function
+    window = spectral_window(nfreq, type="Blackman-Harris", periodic=True)
+    norm_factor = np.sqrt(nfreq / np.sum(window**2.0))
+    window = window * norm_factor
+    window_expand = np.repeat(
+        np.repeat(window[:, np.newaxis, np.newaxis], nbl, axis=1), n_pol, axis=2
+    )
+    data = data * window_expand
+
+    # FFT along freq axis
+    spectra = np.fft.fftshift(np.fft.fft(data, axis=0), axes=0)
+    del data, window
+
+    # Cut at the horizon
+    tau_cut = 1.0
+
+    # Calculate upper and lower delay limits for each baseline
+    lower_limit = nfreq * (0.5 - bb * tau_cut * freq_res)
+    upper_limit = nfreq * (0.5 + bb * tau_cut * freq_res)
+
+    for freq_i in range(nfreq):
+        mask_high = freq_i / upper_limit
+        mask_high_inds = np.nonzero(mask_high > 1.0)
+        if mask_high_inds[0].size > 0:
+            spectra[freq_i, mask_high_inds] = 0
+
+        mask_low = freq_i / lower_limit
+        mask_low_inds = np.nonzero(mask_low < 1.0)
+        if mask_low_inds[0].size > 0:
+            spectra[freq_i, mask_low_inds] = 0
+
+    masked_data = np.fft.ifft(np.fft.fftshift(spectra, axes=0), axis=0)
+    masked_data = masked_data / window_expand
+
+    # Unphase from zenith and cut to the desired band
+    masked_data *= 1.0 / rephase_vals
+    masked_data = masked_data.transpose(2, 0, 1)
+    freq_ind_min = nfreq // 4
+    freq_ind_max = 3 * nfreq // 4
+    vis_model_arr = vis_model_arr[:, freq_ind_min:freq_ind_max]
+    vis_model_arr[:, :, bi_use] = masked_data[:, freq_ind_min:freq_ind_max]
+
+    return vis_model_arr
